add component ListComment, Footer. Modified components, router, and views

+++ src/components/Comment.vue
@@ -2,20 +2,40 @@
   <v-container fluid grid-list-md>
     <v-card>
       <v-card-title>COMMENT</v-card-title>
-      <!-- <v-textarea
-        box
-        auto-grow
-        v-model="comment"
-      >
-      </v-textarea> -->
       <wysiwyg v-model="comment" />
-      <v-btn color="orange" @click="submitComment">Send</v-btn>
+      <v-flex offset-lg10>
+        <v-btn color="orange" @click="submitComment">
+          <v-icon color="white">fa fa-paper-plane-o</v-icon>
+        </v-btn>
+      </v-flex>
       <v-card-text v-if="this.comments">
         <div v-for="(dataComment, index) in comments" :key="index">
-          {{ dataComment.user.email }}: <div v-html="dataComment.comment"></div>
-          <v-btn  @click="deleteComment(index)">DELETE</v-btn>
+          <v-card>
+            <v-card-title>
+              <div>
+                <v-icon dark>account_circle</v-icon>
+                {{ dataComment.user.email }}
+              </div>
+            </v-card-title>
+            <v-card-text>
+              <div v-html="dataComment.comment"></div>
+            </v-card-text>
+            <v-card-actions>
+              <v-flex offset-lg10 v-if="dataComment.user.email === emailFromToken">
+                <v-btn  @click="deleteComment(index)">
+                  <v-icon>fa fa-trash-o</v-icon>
+                </v-btn>
+              </v-flex>
+            </v-card-actions>
+          </v-card>
+          <br>
         </div>
       </v-card-text>
+      <v-flex style="text-align: left">
+        <v-btn  :to="{ path: `/@${author}/${articleId}`, params: { id: articleId } }" style="color: black; text-decoration: none;">
+          <v-icon>fa fa-chevron-left</v-icon>
+        </v-btn>
+      </v-flex>
     </v-card>
   </v-container>
 </template>
@@ -23,19 +43,30 @@
 <script>
 import axios from 'axios';
 import swal from 'sweetalert';
+import jwt from 'jsonwebtoken';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 export default {
+  props: ['id'],
   data: () => ({
     comment: '',
     comments: '',
     token: localStorage.getItem('token'),
     articleId: localStorage.getItem('articleId'),
+    author: localStorage.getItem('author'),
+    url: 'http://localhost:3030/comments/',
+    emailFromToken: '',
   }),
   methods: {
+    getEmailFromToken() {
+      this.emailFromToken = jwt.verify(this.token, process.env.VUE_APP_JWT_SECRET_KEY).email
+    },
     submitComment() {
       axios({
         method: 'POST',
-        url: 'http://localhost:3030/comments',
+        url: this.url,
         data: {
           comment: this.comment,
           articleId: this.articleId,
@@ -46,7 +77,7 @@ export default {
       })
         .then(({ data }) => {
           swal('Add comment success', '', 'success');
-          this.comments.push(data.result)
+          this.comments.push(data.result);
         })
         .catch((err) => {
           swal('Add comment failed', '', 'error');
@@ -55,40 +86,52 @@ export default {
     getComment() {
       axios({
         method: 'GET',
-        url: `http://localhost:3030/comments/${this.articleId}`
+        url: this.url + this.articleId,
       })
         .then(({ data }) => {
-          // console.log('get comment ==>', data)
-          this.comments = data.result
+          this.comments = data.result;
         })
         .catch((err) => {
-          console.log(err)
+          console.log(err);
         })
     },
     deleteComment(index) {
       let id = this.comments[index]._id
-      console.log('ini id yang mau di delete ==>', id)
-      axios({
-        method: 'DELETE',
-        url: `http://localhost:3030/comments/${id}`,
-        headers: {
-          Authorization: this.token,
-        },
-      })
-        .then(() => {
-          swal('Delete comment success', '', 'success');
-          this.comments.splice(index, 1)
+      let commentByUser = this.comments[index].user.email
+      const decoded = jwt.verify(this.token, process.env.VUE_APP_JWT_SECRET_KEY).email
+
+      if (commentByUser === decoded) {
+        axios({
+          method: 'DELETE',
+          url: this.url + id,
+          headers: {
+            Authorization: this.token,
+          },
         })
-        .catch(() => {
-          swal('Delete comment failed', 'Try again!', 'error')
-        })
+          .then(() => {
+            swal('Delete comment success', '', 'success');
+            this.comments.splice(index, 1);
+          })
+          .catch(() => {
+            swal('Delete comment failed', 'Try again!', 'error');
+          });
+      } else {
+        swal('Delete failed', 'You do not have access to delete this comment!', 'warning')
+      }
     },
   },
   created() {
-    this.getComment()
+    this.getComment();
+    this.getEmailFromToken();
   },
   mounted() {
-    this.getComment()
+    this.getComment();
+    this.getEmailFromToken();
+  },
+  watch: {
+    id() {
+      this.getComment();
+    }
   }
 };
 </script>
